chore(bta): update
[1] ────
Update method 'initialize' to be of type 'private' for internal one-off triggering contract initialization, that was kept from 'proxy' migration.
[2] ────
Remove all instances of 'events' for mainnet preparation as a means of code clean. WARNING: will be hard to debug issues in transactions and understand the type of transaction that was triggered (ie: sell/buy).