feat(bta): update
[1] ────
Update method 'calculateBitarenaPriceInStableCoinV2(..)' for dynamic condition of respective check of token ordering in target V3-Pool, which inflicts on price calculation as observed in prior testing on Sepolia for UniswapV3 transaction flows.
[2] ────
Update method 'transferBuySellTakeFees(..)' to be taking action AFTER the official UniswapV3 swap takes place. MOTIVATION: this is done to comply with the UniswapV3 equal amounts passed from one address to the user. Apply respective modifications to make it work as intended, making it an independent method that returns 'void'.